requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.5
selenium==4.15.2
google-generativeai==0.3.0
imapclient==3.0.1
